Edit index, script, style - see changes below
index.html now creates the box div itself.

style.css used to change the attributes of tile and box classes

Edit style.cs to create tiles based on the given width stored in the pixel
variable.

Get width make sure to get the % of the width and height needed to fill
the div box with even squares.